San Diego companies with more than 50 employees
CSV file of about 2200 companies in San Diego with more than 50 employees.   This is the format we will be using for import to the directory.   Note that in includes and "employees" field for estimated number of employees at that location and a "revenue" field for estimated revenue / sales volume.  In addition, each record is categorized by two classification systems - SIC and NAICS and their descriptions.  Please do not alter the content of these fields yet - even if abbreviated - they will be used as the primary lookup for the directory.   This file has been cleaned up to remove retail chains and most branch offices.